src/assessor/requirements.txt
strands-agents==1.36.0
boto3==1.38.27
